+++ packages/admin-rte/src/core/Controls/FeaturesButtonGroup.tsx
@@ -1,7 +1,10 @@
-import { WithStyles } from "@material-ui/core";
+import { ListItemIcon, Menu, MenuItem, WithStyles } from "@material-ui/core";
 import Tooltip from "@material-ui/core/Tooltip";
+import { MoreHoriz } from "@material-ui/icons";
 import { createStyles, withStyles } from "@material-ui/styles";
+import { Editor } from "draft-js";
 import * as React from "react";
+import { FormattedMessage } from "react-intl";
 
 import { IFeatureConfig } from "../types";
 import ControlButton from "./ControlButton";
@@ -9,16 +12,37 @@ import ControlButton from "./ControlButton";
 interface IProps {
     features: IFeatureConfig[];
     disabled?: boolean;
+    editorRef: React.RefObject<Editor>;
+    maxVisible?: number;
 }
 
-function FeaturesButtonGroup({ features, disabled: globallyDisabled, classes }: IProps & WithStyles<typeof styles>) {
+function FeaturesButtonGroup({ features, disabled: globallyDisabled, classes, editorRef, maxVisible }: IProps & WithStyles<typeof styles>) {
+    const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+
+    const handleMoreOptionsClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+        setAnchorEl(event.currentTarget);
+    };
+
+    const handleMenuClose = () => {
+        setAnchorEl(null);
+
+        setTimeout(() => {
+            if (editorRef.current) {
+                editorRef.current.focus();
+            }
+        }, 0);
+    };
+
     if (!features.length) {
         return null;
     }
 
+    const visibleFeatures = maxVisible === undefined ? features : features.slice(0, maxVisible);
+    const additionalFeatures = maxVisible === undefined ? [] : features.slice(maxVisible);
+
     return (
         <div className={classes.root}>
-            {features.map(({ name, onButtonClick, tooltipText, disabled, ...rest }) => (
+            {visibleFeatures.map(({ name, onButtonClick, tooltipText, disabled, ...rest }) => (
                 <div className={classes.buttonWrapper} key={name}>
                     {tooltipText ? (
                         <Tooltip title={tooltipText} placement="top">
@@ -31,11 +55,51 @@ function FeaturesButtonGroup({ features, disabled: globallyDisabled, classes }: 
                     )}
                 </div>
             ))}
+            {additionalFeatures.length > 0 && (
+                <>
+                    <div className={classes.buttonWrapper}>
+                        <Tooltip
+                            title={<FormattedMessage id="cometAdmin.rte.controls.moreOptionsTooltip" defaultMessage="More options" />}
+                            placement="top"
+                        >
+                            <span>
+                                <ControlButton onButtonClick={handleMoreOptionsClick} disabled={globallyDisabled} icon={MoreHoriz} />
+                            </span>
+                        </Tooltip>
+                    </div>
+                    <Menu open={Boolean(anchorEl)} anchorEl={anchorEl} onClose={handleMenuClose}>
+                        {additionalFeatures.map(({ name, onButtonClick, disabled, selected, label, icon: Icon }) => (
+                            <MenuItem
+                                key={name}
+                                onMouseDown={(event) => {
+                                    handleMenuClose();
+
+                                    // See https://reactjs.org/docs/legacy-event-pooling.html for more information.
+                                    event.persist();
+                                    setTimeout(() => {
+                                        onButtonClick?.(event);
+                                    }, 0);
+                                }}
+                                disabled={globallyDisabled || disabled}
+                                selected={selected}
+                                className={classes.listItem}
+                            >
+                                {label}
+                                {Icon && (
+                                    <ListItemIcon className={classes.listItemIcon}>
+                                        <Icon />
+                                    </ListItemIcon>
+                                )}
+                            </MenuItem>
+                        ))}
+                    </Menu>
+                </>
+            )}
         </div>
     );
 }
 
-export type RteFeaturesButtonGroupClassKey = "root" | "buttonWrapper";
+export type RteFeaturesButtonGroupClassKey = "root" | "buttonWrapper" | "listItem" | "listItemIcon";
 
 const styles = () => {
     return createStyles<RteFeaturesButtonGroupClassKey, IProps>({
@@ -48,6 +112,12 @@ const styles = () => {
             "&:last-child": {
                 marginRight: 0,
             },
+        },
+        listItem: {
+            justifyContent: "space-between",
+        },
+        listItemIcon: {
+            justifyContent: "flex-end",
         },
     });
 };